Indepth API

+++ app/src/main/java/Recipe_Logging/FoodData.java
@@ -10,13 +10,13 @@ public class FoodData {
 
     private final String name;
     private final int calories;
-    private final int protein;
-    private final int fats;
-    private final int carbs;
+    private final float protein;
+    private final float fats;
+    private final float carbs;
     private final String servingSize;
 
     //Food info to get
-    public FoodData(String name, int calories, int protein, int fats, int carbs, String servingSize) {
+    public FoodData(String name, int calories, float protein, float fats, float carbs, String servingSize) {
         this.name = name;
         this.calories = calories;
         this.protein = protein;
@@ -27,9 +27,9 @@ public class FoodData {
 
     public String getName() { return name; }
     public int getCalories() { return calories; }  // keep as an INT value
-    public int getProtein() { return protein; }  //change to _g
-    public int getFats() { return fats; }    //Change to _g
-    public int getCarbs() { return carbs; }   //Change to _g
+    public float getProtein() { return protein; }  //change to _g
+    public float getFats() { return fats; }    //Change to _g
+    public float getCarbs() { return carbs; }   //Change to _g
     public String getServingSize() { return servingSize; }
 }
